requirements.txt
opencv-python
pillow
imagehash